feat: implement cat, refactor code
