+++ src/operations/nwd/cd.js
@@ -1,9 +1,9 @@
-import {readdir} from 'fs/promises';
+import {access} from 'fs/promises';
 import {resolve} from 'path';
-import globalVar from '../globalVar.js';
+import globalVar from '../../globalVar.js';
 
 export default async function cd(path) {
   let newPath = resolve(globalVar.currentWorkingDir, path);
-  await readdir(newPath);
+  await access(newPath);
   globalVar.currentWorkingDir = resolve(globalVar.currentWorkingDir, path);
 }
