+++ src/operations/nwd/up.js
@@ -1,5 +1,5 @@
 import {join} from 'path';
-import globalVar from '../globalVar.js';
+import globalVar from '../../globalVar.js';
 
 export default function up() {
   globalVar.currentWorkingDir = join(globalVar.currentWorkingDir, '..');
